Match touch, wireless, graphics and battle communication [skip ci]

+++ include/unk_0201956C.h
@@ -13,12 +13,12 @@ void sub_02019668(UnkStruct_0201956C *a0, int a1, u16 *a2);
 void sub_020196B8(UnkStruct_0201956C *a0, int a1, NARC *narc, int fileId, int a4);
 void sub_020196E8(UnkStruct_0201956C *a0, int a1, int a2, int a3);
 void sub_0201980C(UnkStruct_0201956C *a0, int a1);
-void sub_020198FC(UnkStruct_0201956C *a0, int a1, int a2, int a3, int a4);
+void sub_020198FC(UnkStruct_0201956C *a0, int a1, int a2, int a3, u8 a4);
 void sub_02019934(UnkStruct_0201956C *a0);
 BOOL sub_020199E4(UnkStruct_0201956C *a0, int a1);
 void sub_020199F4(UnkStruct_0201956C *a0, int a1, u8 x, u8 y, u8 width, u8 height, u8 palette);
 u16 *sub_02019B08(UnkStruct_0201956C *a0, int a1);
 void sub_02019B44(UnkStruct_0201956C *a0, int a1, u16 *a2, u16 *a3);
-void sub_02019B70(UnkStruct_0201956C *a0, int a1, int a2, int a3, int a4, int a5);
+void sub_02019B70(UnkStruct_0201956C *a0, int a1, s8 a2, s8 a3, s8 a4, s8 a5);
 
 #endif // GUARD_POKEHEARTGOLD_UNK_0201956C_H
